Fix for PDI-1330 and PDI-1326
Method synchronization takes a lock on the class object and this caused deadlocks in preview and very infrequent HTTP status page updates.

Replaced with block level synchronization on an object specifically made for the status counters locking.

Note that technically, the synchronization could be lifted off of the read methods.  It might result in a stale number once in a while but the end totals should always be right.